Moves autocompletion unittests to separate file
