+++ test/test_server.py
@@ -218,164 +218,6 @@ def test_workspace_symbols():
     check_return(results[1])
 
 
-def test_comp():
-    def check_return(result_array, checks):
-        assert len(result_array) == checks[0]
-        if checks[0] > 0:
-            assert result_array[0]["label"] == checks[1]
-            assert result_array[0]["detail"] == checks[2]
-            try:
-                assert result_array[0]["insertText"] == checks[3]
-            except KeyError:
-                pass
-
-    def comp_request(file_path, line, char):
-        return write_rpc_request(
-            1,
-            "textDocument/completion",
-            {
-                "textDocument": {"uri": str(file_path)},
-                "position": {"line": line, "character": char},
-            },
-        )
-
-    #
-    string = write_rpc_request(1, "initialize", {"rootPath": str(test_dir)})
-    file_path = test_dir / "test_prog.f08"
-    string += comp_request(file_path, 12, 6)
-    string += comp_request(file_path, 13, 6)
-    string += comp_request(file_path, 17, 24)
-    string += comp_request(file_path, 18, 23)
-    string += comp_request(file_path, 20, 7)
-    string += comp_request(file_path, 21, 20)
-    string += comp_request(file_path, 21, 42)
-    string += comp_request(file_path, 23, 26)
-    file_path = test_dir / "subdir" / "test_submod.F90"
-    string += comp_request(file_path, 30, 12)
-    string += comp_request(file_path, 31, 8)
-    string += comp_request(file_path, 31, 23)
-    string += comp_request(file_path, 35, 12)
-    string += comp_request(file_path, 36, 48)
-    file_path = test_dir / "test_inc.f90"
-    string += comp_request(file_path, 10, 2)
-    file_path = test_dir / "subdir" / "test_inc2.f90"
-    string += comp_request(file_path, 3, 2)
-    file_path = test_dir / "subdir" / "test_abstract.f90"
-    string += comp_request(file_path, 7, 12)
-    file_path = test_dir / "subdir" / "test_free.f90"
-    string += comp_request(file_path, 10, 22)
-    string += comp_request(file_path, 14, 27)
-    string += comp_request(file_path, 28, 15)
-    file_path = test_dir / "subdir" / "test_fixed.f"
-    string += comp_request(file_path, 15, 8)
-    string += comp_request(file_path, 15, 21)
-    file_path = test_dir / "subdir" / "test_select.f90"
-    string += comp_request(file_path, 21, 7)
-    string += comp_request(file_path, 23, 7)
-    string += comp_request(file_path, 25, 7)
-    string += comp_request(file_path, 30, 7)
-    file_path = test_dir / "test_block.f08"
-    string += comp_request(file_path, 2, 2)
-    string += comp_request(file_path, 5, 4)
-    string += comp_request(file_path, 8, 6)
-    file_path = test_dir / "subdir" / "test_generic.f90"
-    string += comp_request(file_path, 14, 10)
-    file_path = test_dir / "subdir" / "test_inherit.f90"
-    string += comp_request(file_path, 10, 11)
-    file_path = test_dir / "subdir" / "test_rename.F90"
-    string += comp_request(file_path, 13, 5)
-    string += comp_request(file_path, 14, 5)
-    file_path = test_dir / "subdir" / "test_vis.f90"
-    string += comp_request(file_path, 8, 10)
-    file_path = test_dir / "test_import.f90"
-    string += comp_request(file_path, 15, 20)
-    file_path = test_dir / "completion" / "test_vis_mod_completion.f90"
-    string += comp_request(file_path, 12, 16)
-    string += comp_request(file_path, 12, 24)
-    errcode, results = run_request(string, ["--use_signature_help"])
-    assert errcode == 0
-
-    string = write_rpc_request(1, "initialize", {"rootPath": str(test_dir)})
-    file_path = test_dir / "test_prog.f08"
-    string += comp_request(file_path, 12, 6)
-    errcode, res = run_request(string)
-    assert errcode == 0
-    results.extend(res[1:])
-
-    #
-    exp_results = (
-        # test_prog.f08
-        [1, "myfun", "DOUBLE PRECISION FUNCTION myfun(n, xval)", "myfun"],
-        [9, "glob_sub", "SUBROUTINE glob_sub(n, xval, yval)", "glob_sub"],
-        [1, "bound_nopass", "SUBROUTINE bound_nopass(a, b)", "bound_nopass"],
-        [1, "bound_pass", "SUBROUTINE bound_pass(arg1)", "bound_pass"],
-        [1, "stretch_vector", "TYPE(scaled_vector)"],
-        [6, "scale", "TYPE(scale_type)"],
-        [2, "n", "INTEGER(4)"],
-        [1, "val", "REAL(8)"],
-        # subdir/test_submod.F90
-        [1, "point", "TYPE"],
-        [1, "distance", "REAL"],
-        [2, "x", "REAL"],
-        [1, "point", "TYPE"],
-        [2, "x", "REAL"],
-        # test_inc.f90
-        [2, "val1", "REAL(8)"],
-        # subdir/test_inc2.f90
-        [2, "val1", "REAL(8)"],
-        # subdir/test_abstract.f90
-        [1, "abs_interface", "SUBROUTINE"],
-        # subdir/test_free.f90
-        [1, "DIMENSION(:)", "KEYWORD"],
-        [2, "vector_create", "SUBROUTINE"],
-        [3, "INTENT(IN)", "KEYWORD"],
-        # subdir/test_fixed.f90
-        [1, "bob", "CHARACTER*(LEN=200)"],
-        [1, "dave", "CHARACTER*(20)"],
-        # subdir/test_select.f90
-        [2, "a", "REAL(8)"],
-        [2, "a", "COMPLEX(8)"],
-        [1, "n", "INTEGER(4)"],
-        [2, "a", "REAL(8)"],
-        # test_block.f08
-        [9, "READ", "STATEMENT"],
-        [10, "READ", "STATEMENT"],
-        [11, "READ", "STATEMENT"],
-        # subdir/test_generic.f90
-        [
-            4,
-            "my_gen",
-            "SUBROUTINE my_gen(self, a, b)",
-            "my_gen(${1:self}, ${2:a}, ${3:b})",
-        ],
-        # subdir/test_inherit.f90
-        [1, "val", "REAL(8)"],
-        # subdir/test_rename.F90
-        [1, "localname", "INTEGER"],
-        [2, "renamed_var2", "REAL(8)"],
-        # subdir/test_vis.f90
-        [3, "some_type", "TYPE"],
-        # test_import.f90
-        # TODO: this should be 1, mytype2 should not appear in autocomplete
-        # see #5 and #8 on GitHub
-        [2, "mytype", "TYPE"],
-        # completion/test_vis_mod_completion.f90
-        [1, "some_var", "INTEGER"],
-        [3, "length", "INTEGER"],
-        # test_prog.f08, completion without signature_help
-        # returns the entire completion as a snippet
-        [
-            1,
-            "myfun",
-            "DOUBLE PRECISION FUNCTION myfun(n, xval)",
-            "myfun(${1:n}, ${2:xval})",
-        ],
-    )
-    assert len(exp_results) + 1 == len(results)
-    for i in range(len(exp_results)):
-        check_return(results[i + 1], exp_results[i])
-
-
 def test_sig():
     def check_return(results, checks):
         assert results.get("activeParameter", -1) == checks[0]
